Optimise apply_slack() for size and speed.
To apply proper slack, the original algorithm used to prepare a mask and
then apply the mask to obtain the appropriately rounded-off absolute
time the timer expires.

This patch modifies the masking logic to a bit-shift logic, therby
reducing the complexity and number of operations. Thus obtaining a minor
speed-up.

Change-Id: I0d2af7b7cfa1d7c627ff4a6e8b74373b5e74fe19
Signed-off-by: Chinmay V S <chinmay.v.s@pathpartnertech.com>