msm: ultrasound: Add support for general button events
The side buttons bitmap was expanded to allow sending button
events for tool type buttons (stylus, rubber) and also for a new
button which signals that the hovering icon should be displayed.

The event source field was removed as it was previously used to
set the tool type, which is now done explicitly via the buttons
bitmap.

Change-Id: Id08921c392eeec8928ca0c730988f8c83be2ba8f
Signed-off-by: Lior Barenboim <liorb@codeaurora.org>