arm: kernel: Drop warning about return_address not using unwind tables
The warning was introduced in 2009 (implement CALLER_ADDRESSx).
The only "problem" here is that
CALLER_ADDRESSx for x > 1 returns NULL which doesn't do much harm.

The drawback of implementing a fix (i.e. use unwind tables to implement CALLER_ADDRESSx)
is that much of the unwinder code would need to be marked as not traceable.

Change-Id: I6b9bf44ef272006b15f8b1c1435ecd2a07c6258a
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>