vfs: read file_handle only once in handle_to_path
We used to read file_handle twice. Once to get the amount of extra bytes, and
once to fetch the entire structure.

This may be problematic since we do size verifications only after the first
read, so if the number of extra bytes changes in userspace between the first
and second calls, we'll have an incoherent view of file_handle.

Instead, read the constant size once, and copy that over to the final
structure without having to re-read it again.

Change-Id: Ib8f008e79d919a733419d1e040b8c6bbc07724f6
Signed-off-by: Sasha Levin <sasha.levin@oracle.com>