ARM: dts: ms013g: Enable dynamic FPS on suspend/resume
Allows the panel to dynamically change its FPS, saving power
during A/V usecases and idle.

Unfortunately, our MDSS subsystem is quite old so we can't make use
of the newer and more useful dynamic FPS options, allowing for
much greater power savings.

One improvement to make in the future is to update the MDSS subsystem
with the appropriate patches to enable the newer options, or we
could just rebase on a newer CAF base - both choices are equally
tiresome.

Change-Id: Ie4f8760bea13aea483384990686418f38ef7abeb