ms013g: Disable module support
Modules are a security vulnerability in Android and have been disabled
in Nexus devices for a long time. Though not restricted by Google for
known reasons, we should still try our best to reduce the number of
modules. As we no longer have any more modules, let's disable module
support.

Change-Id: I5bc72b0ff11bcb043f8968bc19e2eb18d51cf2b8