block: Add driver to change the I/O scheduler when the screen turns off
This driver changes the I/O scheduler for a specified request queue to
Noop when the screen turns off in order to improve power consumption while
the screen is off, as well as reduce resume latency when the screen is
turned back on. After the screen is turned back on, the driver will change
the I/O scheduler back to what it originally was after a 10000ms delay.

Change-Id: Ifc7484df390c577fc8ef0eeda57ba7f82896e3c4
Signed-off-by: Sultanxda <sultanxda@gmail.com>